+++ packages/flame/lib/src/components/component_set.dart
@@ -30,7 +30,7 @@ class ComponentSet extends QueryableOrderedSet<Component> {
 
   /// This is the "prepare" function that will be called *before* the
   /// component is added to the component list by the add/addAll methods.
-  final void Function(Component child, {BaseGame? gameRef}) prepare;
+  final void Function(Component child) prepare;
 
   ComponentSet(
     int Function(Component e1, Component e2)? compare,
@@ -71,15 +71,15 @@ class ComponentSet extends QueryableOrderedSet<Component> {
   /// You can await for the onLoad function, if present.
   /// This method can be considered sync for all intents and purposes if no
   /// onLoad is provided by the component.
-  Future<void> addChild(Component c, {BaseGame? gameRef}) async {
-    prepare(c, gameRef: gameRef);
+  Future<void> addChild(Component component) async {
+    prepare(component);
 
-    final loadFuture = c.onLoad();
+    final loadFuture = component.onLoad();
     if (loadFuture != null) {
       await loadFuture;
     }
 
-    _addLater.add(c);
+    _addLater.add(component);
   }
 
   /// Prepares and registers a list of component to be added on the next game
@@ -90,7 +90,7 @@ class ComponentSet extends QueryableOrderedSet<Component> {
     Iterable<Component> components, {
     BaseGame? gameRef,
   }) async {
-    final ps = components.map((c) => addChild(c, gameRef: gameRef));
+    final ps = components.map(addChild);
     await Future.wait(ps);
   }
 
@@ -163,7 +163,7 @@ class ComponentSet extends QueryableOrderedSet<Component> {
   ///
   /// You must still provide your [prepare] function depending on the context.
   static ComponentSet createDefault(
-    void Function(Component child, {BaseGame? gameRef}) prepare,
+    void Function(Component child) prepare,
   ) {
     return ComponentSet(
       Comparing.on<Component>((c) => c.priority),