+++ packages/flame/CHANGELOG.md
@@ -46,6 +46,8 @@
  - Added `FocusNode` on the game widget and improved keyboard handling in the game.
  - Added ability to have custom mouse cursor on the `GameWidget` region
  - Change sprite component to default to the Sprite size if not provided
+ - `Game` is now a `Component`
+ - `ComponentEffect` is now a `Component`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps